backend/requirements.txt
asgiref==3.2.10
requests==2.26.0
django==2.2.16
django-filter==2.4.0
django-colorfield==0.8.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.4.1
flake8==5.0.4
gunicorn==20.0.4
isort==5.11.4
psycopg2-binary==2.8.6
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==0.20.0
pytz==2020.1
sqlparse==0.3.1
social-auth-app-django==4.0.0
Pillow==9.3.0
